Moved Graphs to their own container
