+++ client/src/containers/Graphs/GraphList.jsx
@@ -1,0 +1,47 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import { Typography } from '@material-ui/core';
+import ListItem from 'components/ListItem';
+import GraphWrapper from './GraphWrapper';
+
+const useStyles = makeStyles(() => ({
+  content: {
+    height: '350px',
+    paddingTop: '2em',
+    overflow: 'hidden',
+    flexBasis: '50%',
+  },
+}));
+
+const GraphList = ({ stats, unitNames, graphMap }) => {
+  const classes = useStyles();
+  const firstLoad = (!stats.payload || !stats.payload.length) && stats.pending;
+
+  return (
+    <Typography component="div">
+      {[...graphMap].map(([name, Graph]) => (
+        <ListItem
+          key={name}
+          header={name}
+          collapsible
+          loading={stats.pending}
+          loaderDelay={firstLoad ? 0 : 350}
+        >
+          <GraphWrapper
+            loading={(!stats.payload || !stats.payload.length) && stats.pending}
+            numUnits={unitNames.length}
+            error={Boolean(stats.error)}
+          >
+            <Graph
+              className={classes.content}
+              results={stats.payload}
+              unitNames={unitNames}
+            />
+          </GraphWrapper>
+        </ListItem>
+      ))}
+    </Typography>
+  );
+};
+
+export default GraphList;
